Remove using instanceof through use of polymorphism for events + population passed in through constructor
