+++ src/main/java/uk/co/ramp/Outbreak.java
@@ -5,7 +5,8 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.co.ramp.event.EventList;
-import uk.co.ramp.event.EventProcessor;
+import uk.co.ramp.event.EventProcessorRunner;
+import uk.co.ramp.event.processor.InfectionEventProcessor;
 import uk.co.ramp.event.types.Event;
 import uk.co.ramp.event.types.ImmutableInfectionEvent;
 import uk.co.ramp.event.types.InfectionEvent;
@@ -18,7 +19,6 @@ import uk.co.ramp.io.types.DiseaseProperties;
 import uk.co.ramp.io.types.StandardProperties;
 import uk.co.ramp.people.Case;
 import uk.co.ramp.people.VirusStatus;
-import uk.co.ramp.utilities.UtilitiesBean;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -37,34 +37,30 @@ public class Outbreak {
     private final DiseaseProperties diseaseProperties;
     private final EventList eventList;
     private final InitialCaseReader initialCaseReader;
-    private final EventProcessor eventProcessor;
-    private final UtilitiesBean utils;
+    private final EventProcessorRunner eventProcessorRunner;
     private final LogDailyOutput outputLog;
+    private final InfectionEventProcessor infectionEventProcessor;
 
-    private Map<Integer, Case> population;
+    private final Population population;
     private final Map<Integer, CmptRecord> records = new HashMap<>();
     private static final String INFECTION_MAP = "infectionMap.txt";
 
 
     @Autowired
-    public Outbreak(DiseaseProperties diseaseProperties, StandardProperties standardProperties,
-                    UtilitiesBean utils, LogDailyOutput outputLog, InitialCaseReader initialCaseReader,
-                    EventList eventList, EventProcessor eventProcessor) {
+    public Outbreak(Population population, DiseaseProperties diseaseProperties, StandardProperties standardProperties,
+                    LogDailyOutput outputLog, InitialCaseReader initialCaseReader,
+                    EventList eventList, EventProcessorRunner eventProcessorRunner,
+                    InfectionEventProcessor infectionEventProcessor) {
 
+        this.population = population;
         this.initialCaseReader = initialCaseReader;
         this.diseaseProperties = diseaseProperties;
         this.properties = standardProperties;
-        this.utils = utils;
         this.outputLog = outputLog;
-        this.eventProcessor = eventProcessor;
+        this.eventProcessorRunner = eventProcessorRunner;
+        this.infectionEventProcessor = infectionEventProcessor;
         this.eventList = eventList;
-
     }
-
-    public void setPopulation(Map<Integer, Case> population) {
-        this.population = population;
-    }
-
 
     public Map<Integer, CmptRecord> propagate() {
 
@@ -86,6 +82,7 @@ public class Outbreak {
                 nextStatus(EXPOSED).
                 exposedTime(0).
                 id(-1).
+                eventProcessor(infectionEventProcessor).
                 time(0).build();
 
         infectedIds.forEach(id -> virusEvents.add(ImmutableInfectionEvent.copyOf(genericEvent).withId(id)));
@@ -102,7 +99,7 @@ public class Outbreak {
         runContactData(timeLimit, randomInfectionRate);
 
         try (Writer writer = new FileWriter(new File(INFECTION_MAP))) {
-            new InfectionMap(population).outputMap(writer);
+            new InfectionMap(population.view()).outputMap(writer);
             eventList.output();
         } catch (IOException e) {
             String message = "An error occurred generating the infection map";
@@ -123,8 +120,7 @@ public class Outbreak {
 
         for (int time = 0; time <= timeLimit; time++) {
 
-            eventProcessor.setPopulation(population);
-            eventProcessor.process(time, randomInfectionRate, lastContact);
+            eventProcessorRunner.process(time, randomInfectionRate, lastContact);
             updateLogActiveCases(time);
 
             // stop random infections after contacts end
@@ -141,13 +137,13 @@ public class Outbreak {
     }
 
     void updateLogActiveCases(int time) {
-        Map<VirusStatus, Integer> stats = utils.getCmptCounts(population);
+        Map<VirusStatus, Integer> stats = population.getCmptCounts();
         CmptRecord cmptRecord = outputLog.log(time, stats);
         records.put(time, cmptRecord);
     }
 
     private int activeCases() {
-        var stats = utils.getCmptCounts(population);
+        var stats = population.getCmptCounts();
         return stats.get(EXPOSED) + stats.get(ASYMPTOMATIC) + stats.get(PRESYMPTOMATIC) + stats.get(SYMPTOMATIC) + stats.get(SEVERELY_SYMPTOMATIC);
     }
