+++ src/main/java/uk/co/ramp/event/types/Event.java
@@ -4,5 +4,5 @@ public interface Event {
 
     int time();
 
-
+    ProcessedEventResult processEvent();
 }
