+++ src/main/java/uk/co/ramp/event/types/InfectionEvent.java
@@ -1,16 +1,17 @@
 package uk.co.ramp.event.types;
 
 import org.immutables.value.Value;
-import uk.co.ramp.people.Case;
+import uk.co.ramp.event.processor.InfectionEventProcessor;
 
 @Value.Immutable
 public interface InfectionEvent extends CommonVirusEvent {
-
-    default void applyEventToCase(Case aCase) {
-        aCase.setVirusStatus(aCase.virusStatus().transitionTo(this.nextStatus()));
-    }
-
     int exposedBy();
 
     int exposedTime();
+
+    InfectionEventProcessor eventProcessor();
+
+    default ProcessedEventResult processEvent() {
+        return eventProcessor().processEvent(this);
+    }
 }
