+++ src/main/java/uk/co/ramp/event/types/VirusEvent.java
@@ -3,15 +3,15 @@ package uk.co.ramp.event.types;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.immutables.value.Value;
-import uk.co.ramp.people.Case;
+import uk.co.ramp.event.processor.VirusEventProcessor;
 
 @Value.Immutable
 @JsonSerialize
 @JsonDeserialize
 public interface VirusEvent extends CommonVirusEvent {
+    VirusEventProcessor eventProcessor();
 
-    default void applyEventToCase(Case aCase) {
-        aCase.setVirusStatus(aCase.virusStatus().transitionTo(this.nextStatus()));
+    default ProcessedEventResult processEvent() {
+        return eventProcessor().processEvent(this);
     }
-
 }
