+++ src/test/java/uk/co/ramp/event/FormattedEventFactoryTest.java
@@ -4,9 +4,9 @@ import net.bytebuddy.utility.RandomString;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mockito;
 import uk.co.ramp.LogSpy;
 import uk.co.ramp.TestUtils;
+import uk.co.ramp.event.processor.*;
 import uk.co.ramp.event.types.*;
 import uk.co.ramp.people.AlertStatus;
 import uk.co.ramp.people.Case;
@@ -16,6 +16,7 @@ import java.util.Random;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.hamcrest.CoreMatchers.containsString;
+import static org.mockito.Mockito.mock;
 
 public class FormattedEventFactoryTest {
 
@@ -34,7 +35,7 @@ public class FormattedEventFactoryTest {
         VirusStatus old = VirusStatus.EXPOSED;
         VirusStatus next = VirusStatus.PRESYMPTOMATIC;
 
-        VirusEvent virusEvent = ImmutableVirusEvent.builder().time(time).id(id).oldStatus(old).nextStatus(next).build();
+        VirusEvent virusEvent = ImmutableVirusEvent.builder().time(time).id(id).oldStatus(old).nextStatus(next).eventProcessor(mock(VirusEventProcessor.class)).build();
 
         ImmutableFormattedEvent formatted = formattedEventFactory.create(virusEvent);
 
@@ -56,7 +57,8 @@ public class FormattedEventFactoryTest {
         VirusStatus old = VirusStatus.SUSCEPTIBLE;
         VirusStatus next = VirusStatus.EXPOSED;
 
-        InfectionEvent infectionEvent = ImmutableInfectionEvent.builder().time(time).id(id).oldStatus(old).nextStatus(next).exposedBy(exposer).exposedTime(exposedTime).build();
+        InfectionEventProcessor eventProcessor = mock(InfectionEventProcessor.class);
+        InfectionEvent infectionEvent = ImmutableInfectionEvent.builder().time(time).id(id).oldStatus(old).nextStatus(next).exposedBy(exposer).exposedTime(exposedTime).eventProcessor(eventProcessor).build();
 
         ImmutableFormattedEvent formatted = formattedEventFactory.create(infectionEvent);
 
@@ -98,7 +100,9 @@ public class FormattedEventFactoryTest {
         AlertStatus old = AlertStatus.NONE;
         AlertStatus next = AlertStatus.ALERTED;
 
-        AlertEvent alertEvent = ImmutableAlertEvent.builder().time(time).id(id).nextStatus(next).oldStatus(old).build();
+        AlertEventProcessor eventProcessor = mock(AlertEventProcessor.class);
+
+        AlertEvent alertEvent = ImmutableAlertEvent.builder().time(time).id(id).nextStatus(next).oldStatus(old).eventProcessor(eventProcessor).build();
 
         ImmutableFormattedEvent formatted = formattedEventFactory.create(alertEvent);
 
@@ -115,8 +119,9 @@ public class FormattedEventFactoryTest {
         double weight = random.nextDouble() * random.nextInt(100);
         String label = RandomString.make(20);
 
+        ContactEventProcessor eventProcessor = mock(ContactEventProcessor.class);
 
-        ContactEvent contactEvent = ImmutableContactEvent.builder().time(time).to(id).from(id2).weight(weight).label(label).build();
+        ContactEvent contactEvent = ImmutableContactEvent.builder().time(time).to(id).from(id2).weight(weight).label(label).eventProcessor(eventProcessor).build();
 
         ImmutableFormattedEvent formatted = formattedEventFactory.create(contactEvent);
 
@@ -149,12 +154,12 @@ public class FormattedEventFactoryTest {
         VirusStatus nextVirus = VirusStatus.EXPOSED;
 
 
-        Event contactEvent = ImmutableContactEvent.builder().time(time).to(id).from(id2).weight(weight).label(label).build();
+        Event contactEvent = ImmutableContactEvent.builder().time(time).to(id).from(id2).weight(weight).label(label).eventProcessor(mock(ContactEventProcessor.class)).build();
         Event policyEvent = ImmutablePolicyEvent.builder().time(time).build();
-        Event alertEvent = ImmutableAlertEvent.builder().time(time).id(id).nextStatus(nextAlert).oldStatus(oldAlert).build();
-        Event infectionEvent = ImmutableInfectionEvent.builder().time(time).id(id).oldStatus(oldVirus).nextStatus(nextVirus).exposedBy(exposer).exposedTime(exposedTime).build();
-        Event virusEvent = ImmutableVirusEvent.builder().time(time).id(id).oldStatus(oldVirus).nextStatus(nextVirus).build();
-        Event notCovered = Mockito.mock(CommonVirusEvent.class);
+        Event alertEvent = ImmutableAlertEvent.builder().time(time).id(id).nextStatus(nextAlert).oldStatus(oldAlert).eventProcessor(mock(AlertEventProcessor.class)).build();
+        Event infectionEvent = ImmutableInfectionEvent.builder().time(time).id(id).oldStatus(oldVirus).nextStatus(nextVirus).exposedBy(exposer).exposedTime(exposedTime).eventProcessor(mock(InfectionEventProcessor.class)).build();
+        Event virusEvent = ImmutableVirusEvent.builder().time(time).id(id).oldStatus(oldVirus).nextStatus(nextVirus).eventProcessor(mock(VirusEventProcessor.class)).build();
+        Event notCovered = mock(CommonVirusEvent.class);
 
         // TODO: not printing these as they're an input... should discuss with group.
         ImmutableFormattedEvent event = formattedEventFactory.create(contactEvent);
