+++ src/test/resources/logback-test.xml
@@ -2,7 +2,8 @@
 <configuration>
     <include resource="org/springframework/boot/logging/logback/base.xml"/>
     <logger name="uk.co.ramp.Outbreak" level="debug"/>
-    <logger name="uk.co.ramp.event.EventProcessor" level="debug"/>
+    <logger name="uk.co.ramp.event.EventProcessorRunner" level="debug"/>
+    <logger name="uk.co.ramp.event.processor.ContactEventProcessor" level="debug"/>
     <logger name="org.springframework.core " level="error"/>
     <logger name="org.springframework.beans" level="error"/>
     <logger name="org.springframework.context" level="error"/>
